+++ src/compiler/scala/tools/nsc/typechecker/Typers.scala
@@ -1259,7 +1259,7 @@ trait Typers { self: Analyzer =>
      */
     def addGetterSetter(stat: Tree): List[Tree] = stat match {
       case ValDef(mods, name, tpt, rhs)
-        if (mods.flags & (PRIVATE | LOCAL)) != (PRIVATE | LOCAL) && !stat.symbol.isModuleVar =>
+        if (mods.flags & (PRIVATE | LOCAL)) != (PRIVATE | LOCAL).toLong && !stat.symbol.isModuleVar =>
         val isDeferred = mods hasFlag DEFERRED
         val value = stat.symbol
         val getter = if (isDeferred) value else value.getter(value.owner)
